+++ src/elements/governance_stage_calls.rs
@@ -1,15 +1,13 @@
 use super::{
     call_list::{Call, CallList},
+    deployed_addresses::DeployedAddresses,
     fixed_force_deployment::FixedForceDeploymentsData,
-    protocol_version::ProtocolVersion,
+    
+    set_new_version_upgrade::{self, setNewVersionUpgradeCall},
 };
 use crate::{
     elements::initialize_data_new_chain::InitializeDataNewChain,
-    get_expected_old_protocol_version,
-    utils::{
-        compute_selector,
-        facet_cut_set::{self, FacetCutSet, FacetInfo},
-    },
+    utils::facet_cut_set::{self, FacetCutSet, FacetInfo},
     verifiers::Verifiers,
 };
 use alloy::{
@@ -18,7 +16,15 @@ use alloy::{
     sol,
     sol_types::{SolCall, SolValue},
 };
+use anyhow::Context;
 
+pub struct GovernanceStage0Calls {
+    pub calls: CallList,
+}
+
+pub struct GovernanceStage1Calls {
+    pub calls: CallList,
+}
 pub struct GovernanceStage2Calls {
     pub calls: CallList,
 }
@@ -77,7 +83,7 @@ sol! {
     function facets() external view returns (Facet[] memory result);
 }
 
-impl GovernanceStage2Calls {
+impl GovernanceStage1Calls {
     /// Verifies an upgrade call by decoding its data and comparing the proxy and implementation addresses.
     pub fn verify_upgrade_call(
         &self,
@@ -118,35 +124,48 @@ impl GovernanceStage2Calls {
         Ok(())
     }
 
-    /// Verifies all the governance stage 2 calls.
+    /// Verifies all the governance stage 1 calls.
     /// Returns a pair of expected diamond cut data as well as expected fixed force deployments data.
     pub async fn verify(
         &self,
         verifiers: &crate::verifiers::Verifiers,
         result: &mut crate::verifiers::VerificationResult,
         expected_chain_creation_facets: FacetCutSet,
+        deployed_addresses: &DeployedAddresses,
+        expected_upgrade_facets: FacetCutSet,
+        expected_chain_upgrade_diamond_cut: &str,
     ) -> anyhow::Result<(String, String)> {
-        result.print_info("== Gov stage 2 calls ===");
+        result.print_info("== Gov stage 1 calls ===");
+
+        // Stage1 is where most of the upgrade happens.
+        // It usually consists of 3 parts:
+        // * upgrading proxies (we deploy a new implementation and point existing proxy to it)
+        // * upgrading chain creation parameters (telling the system how the new chains should look like)
+        // * saving the information on how to upgrade existing chains (set new version upgrade)
+
+        // Optionally for some upgrades we might have additional contract calls
+        // (for example when we added a new type of bridge, we also included a call to bridgehub to set its address etc)
+
 
         let list_of_calls = [
+            // Proxy upgrades
             ("transparent_proxy_admin", "upgrade(address,address)"),
-            ("transparent_proxy_admin", "upgradeAndCall(address,address,bytes)"),
             ("transparent_proxy_admin", "upgrade(address,address)"),
             ("transparent_proxy_admin", "upgrade(address,address)"),
+            ("transparent_proxy_admin", "upgrade(address,address)"),
+            ("transparent_proxy_admin", "upgrade(address,address)"),
+            // index = 5
             (
                 "state_transition_manager",
                 "setChainCreationParams((address,bytes32,uint64,bytes32,((address,uint8,bool,bytes4[])[],address,bytes),bytes))",
             ),
-            ("bridgehub_proxy", "setAddresses(address,address,address)"),
-            ("old_shared_bridge_proxy", "setL1NativeTokenVault(address)"),
-            ("old_shared_bridge_proxy", "setL1AssetRouter(address)"),
-            ("state_transition_manager", "setProtocolVersionDeadline(uint256,uint256)"),
-            ("upgrade_timer", "checkDeadline()"),
-            /*(
-                "protocol_upgrade_handler_transparent_proxy_admin",
-                "upgradeAndCall(address,address,bytes)",
-            ),*/
+
+            ("state_transition_manager",
+            "setNewVersionUpgrade(((address,uint8,bool,bytes4[])[],address,bytes),uint256,uint256,uint256)"),
         ];
+        const SET_NEW_VERSION_INDEX: usize = 6;
+        const SET_CHAIN_CREATION_INDEX: usize = 5;
+
 
         self.calls.verify(&list_of_calls, verifiers, result)?;
 
@@ -160,22 +179,20 @@ impl GovernanceStage2Calls {
             None,
         )?;
 
-        // Compute the selector once so that its lifetime is extended.
-        let init_v2_selector = compute_selector("initializeV2()");
         self.verify_upgrade_call(
             verifiers,
             result,
             &self.calls.elems[1],
             "bridgehub_proxy",
             "bridgehub_implementation_addr",
-            Some(init_v2_selector.as_str()),
+            None,
         )?;
 
         self.verify_upgrade_call(
             verifiers,
             result,
             &self.calls.elems[2],
-            "old_shared_bridge_proxy",
+            "l1_nullifier_proxy_addr",
             "l1_nullifier_implementation_addr",
             None,
         )?;
@@ -184,14 +201,73 @@ impl GovernanceStage2Calls {
             verifiers,
             result,
             &self.calls.elems[3],
-            "legacy_erc20_bridge_proxy",
-            "erc20_bridge_implementation_addr",
+            "l1_asset_router_proxy",
+            "l1_asset_router_implementation_addr",
             None,
         )?;
+        self.verify_upgrade_call(
+            verifiers,
+            result,
+            &self.calls.elems[4],
+            "native_token_vault",
+            "native_token_vault_implementation_addr",
+            None,
+        )?;
+        
+
+        // Verify setNewVersionUpgrade
+        {
+            let calldata = &self.calls.elems[SET_NEW_VERSION_INDEX].data;
+            let data = setNewVersionUpgradeCall::abi_decode(calldata, true).unwrap();
+
+            if data.oldProtocolVersionDeadline != U256::MAX {
+                result.report_error("Wrong old protocol version deadline for stage1 call");
+            }
+
+            let diamond_cut = data.diamondCut;
+            if alloy::hex::encode(diamond_cut.abi_encode())
+                != expected_chain_upgrade_diamond_cut[2..]
+            {
+                result.report_error(&format!(
+                    "Invalid chain upgrade diamond cut. Expected: {}\n Received: {}",
+                    expected_chain_upgrade_diamond_cut,
+                    alloy::hex::encode(diamond_cut.abi_encode())
+                ));
+            }
+            // TODO: verify diamond_cut.initAddress
+
+            // should match state_transiton.default_upgrade
+            println!("diamond_cut.initAddress: {}", diamond_cut.initAddress);
+
+            /*if diamond_cut.initAddress != deployed_addresses.l1_gateway_upgrade {
+                result.report_error(&format!(
+                    "Unexpected init address for the diamond cut: {}, expected {}",
+                    diamond_cut.initAddress, deployed_addresses.l1_gateway_upgrade
+                ));
+            }*/
+
+            verity_facet_cuts(&diamond_cut.facetCuts, result, expected_upgrade_facets).await;
+
+            let upgrade = crate::elements::set_new_version_upgrade::upgradeCall::abi_decode(
+                &diamond_cut.initCalldata,
+                true,
+            )
+            .unwrap();
+
+            upgrade
+                ._proposedUpgrade
+                .verify(
+                    verifiers,
+                    result,
+                    deployed_addresses.l1_bytecodes_supplier_addr,
+                )
+                .await
+                .context("proposed upgrade")?;
+        }
 
         // Verify setChainCreationParams call.
         let (chain_creation_diamond_cut, force_deployments) = {
-            let decoded = setChainCreationParamsCall::abi_decode(&self.calls.elems[4].data, true)
+            let decoded = setChainCreationParamsCall::abi_decode(&self.calls.elems[SET_CHAIN_CREATION_INDEX].data, true)
                 .expect("Failed to decode setChainCreationParams call");
             decoded
                 ._chainCreationParams
@@ -209,65 +285,6 @@ impl GovernanceStage2Calls {
                 hex::encode(forceDeploymentsData),
             )
         };
-
-        // Verify setAddresses call.
-        {
-            let decoded = setAddressesCall::abi_decode(&self.calls.elems[5].data, true)
-                .expect("Failed to decode setAddresses call");
-            result.expect_address(verifiers, &decoded._assetRouter, "l1_asset_router_proxy");
-            result.expect_address(verifiers, &decoded._l1CtmDeployer, "ctm_deployment_tracker");
-            result.expect_address(verifiers, &decoded._messageRoot, "l1_message_root");
-        }
-
-        // Verify setL1NativeTokenVault call.
-        {
-            let decoded = setL1NativeTokenVaultCall::abi_decode(&self.calls.elems[6].data, true)
-                .expect("Failed to decode setL1NativeTokenVault call");
-            result.expect_address(
-                verifiers,
-                &decoded._l1NativeTokenVault,
-                "native_token_vault",
-            );
-        }
-
-        // Verify setL1AssetRouter call.
-        {
-            let decoded = setL1AssetRouterCall::abi_decode(&self.calls.elems[7].data, true)
-                .expect("Failed to decode setL1AssetRouter call");
-            result.expect_address(verifiers, &decoded._l1AssetRouter, "l1_asset_router_proxy");
-        }
-
-        // Verify setProtocolVersionDeadline call.
-        {
-            let decoded =
-                setProtocolVersionDeadlineCall::abi_decode(&self.calls.elems[8].data, true)
-                    .expect("Failed to decode setProtocolVersionDeadline call");
-            let pv = ProtocolVersion::from(decoded.protocolVersion);
-            let expected_old = get_expected_old_protocol_version();
-
-            if pv != expected_old {
-                result.report_warn(&format!(
-                    "Invalid protocol version: {} - expected {}",
-                    pv, expected_old
-                ));
-            }
-
-            if decoded.newDeadline != U256::ZERO {
-                result.report_error(&format!(
-                    "Expected deadline to be 0, but got {}",
-                    decoded.newDeadline
-                ));
-            }
-        }
-
-        /*self.verify_upgrade_call(
-            verifiers,
-            result,
-            &self.calls.elems[10],
-            "protocol_upgrade_handler_proxy",
-            "new_protocol_upgrade_handler_impl",
-            Some(""),
-        )?;*/
 
         Ok((chain_creation_diamond_cut, force_deployments))
     }
@@ -326,12 +343,16 @@ impl ChainCreationParams {
         )
         .await?;
 
-        let fixed_force_deployments_data =
-            FixedForceDeploymentsData::abi_decode(&self.forceDeploymentsData, true)
-                .expect("Failed to decode FixedForceDeploymentsData");
-        fixed_force_deployments_data
-            .verify(verifiers, result)
-            .await?;
+        if self.forceDeploymentsData.is_empty() {
+            result.report_error("Force deployments data is empty");
+        } else {
+            let fixed_force_deployments_data =
+                FixedForceDeploymentsData::abi_decode(&self.forceDeploymentsData, true)
+                    .expect("Failed to decode FixedForceDeploymentsData");
+            fixed_force_deployments_data
+                .verify(verifiers, result)
+                .await?;
+        }
 
         Ok(())
     }
@@ -383,4 +404,90 @@ pub async fn verify_chain_creation_diamond_cut(
     initialize_data_new_chain.verify(verifiers, result).await?;
 
     Ok(())
+}
+
+
+
+pub async fn verity_facet_cuts(
+    facet_cuts: &[set_new_version_upgrade::FacetCut],
+    result: &mut crate::verifiers::VerificationResult,
+    expected_upgrade_facets: FacetCutSet,
+) {
+    // We ensure two invariants here:
+    // - Firstly we use `Remove` operations only. This is mainly for ensuring that
+    // the upgrade will pass.
+    // - Secondly, we ensure that the set of operations is identical.
+    let mut used_add = false;
+    let mut proposed_facet_cuts = FacetCutSet::new();
+    facet_cuts.iter().for_each(|facet| {
+        let action = match facet.action {
+            set_new_version_upgrade::Action::Add => {
+                used_add = true;
+                facet_cut_set::Action::Add
+            }
+            set_new_version_upgrade::Action::Remove => {
+                assert!(!used_add, "Unexpected `Remove` operation after `Add`");
+                facet_cut_set::Action::Remove
+            }
+            set_new_version_upgrade::Action::Replace => panic!("Replace unexpected"),
+            set_new_version_upgrade::Action::__Invalid => panic!("Invalid unexpected"),
+        };
+
+        proposed_facet_cuts.add_facet(FacetInfo {
+            facet: facet.facet,
+            action,
+            is_freezable: facet.isFreezable,
+            selectors: facet.selectors.iter().map(|x| x.0).collect(),
+        });
+    });
+
+    if proposed_facet_cuts != expected_upgrade_facets {
+        result.report_error(&format!(
+            "Incorrect facet cuts. Expected {:#?}\nReceived: {:#?}",
+            expected_upgrade_facets, proposed_facet_cuts
+        ));
+    }
+}
+
+impl GovernanceStage0Calls {
+    /// Stage0 is executed before the main upgrade even starts.
+    pub(crate) async fn verify(
+        &self,
+        verifiers: &crate::verifiers::Verifiers,
+        result: &mut crate::verifiers::VerificationResult,
+    ) -> anyhow::Result<()> {
+        result.print_info("== Gov stage 0 calls ===");
+
+        let list_of_calls = [
+            ("bridgehub_proxy", "pauseMigration()"),
+        ];
+        // If this is just a single call without any params, we don't have to check
+        // anything else.
+
+        self.calls
+            .verify(&list_of_calls, verifiers, result)
+            .context("calls")?;
+        Ok(())
+    }
+}
+impl GovernanceStage2Calls {
+    /// Stage2 is executed after all the chains have upgraded.
+    pub(crate) async fn verify(
+        &self,
+        verifiers: &crate::verifiers::Verifiers,
+        result: &mut crate::verifiers::VerificationResult,
+    ) -> anyhow::Result<()> {
+        result.print_info("== Gov stage 2 calls ===");
+
+        let list_of_calls = [
+            ("bridgehub_proxy", "unpauseMigration()"),
+        ];
+        // If this is just a single call without any params, we don't have to check
+        // anything else.
+
+        self.calls
+            .verify(&list_of_calls, verifiers, result)
+            .context("calls")?;
+        Ok(())
+    }
 }