+++ src/elements/set_new_version_upgrade.rs
@@ -3,11 +3,18 @@ use std::collections::HashSet;
 use alloy::{
     primitives::{Address, FixedBytes, U256},
     sol,
+    sol_types::SolCall,
 };
+use anyhow::Context;
 
 use crate::get_expected_new_protocol_version;
 
-use super::{post_upgrade_calldata::PostUpgradeCalldata, protocol_version::ProtocolVersion};
+use super::{
+    force_deployment::{
+        expected_force_deployments, forceDeployOnAddressesCall, verify_force_deployments,
+    },
+    protocol_version::ProtocolVersion,
+};
 
 const DEPLOYER_SYSTEM_CONTRACT: u32 = 0x8006;
 const FORCE_DEPLOYER_ADDRESS: u32 = 0x8007;
@@ -104,7 +111,7 @@ sol! {
 
 impl upgradeCall {} // Placeholder implementation.
 
-const EXPECTED_BYTECODES: [&str; 46] = [
+const EXPECTED_BYTECODES: [&str; 42] = [
     "CodeOracle.yul",
     "EcAdd.yul",
     "EcMul.yul",
@@ -118,7 +125,6 @@ const EXPECTED_BYTECODES: [&str; 46] = [
     "EvmEmulator.yul",
     "Identity.yul",
     "EvmGasManager.yul",
-    "l1-contracts/BeaconProxy",
     "l1-contracts/BridgedStandardERC20",
     "l1-contracts/Bridgehub",
     "l1-contracts/L2AssetRouter",
@@ -126,7 +132,6 @@ const EXPECTED_BYTECODES: [&str; 46] = [
     "l1-contracts/L2SharedBridgeLegacy",
     "l1-contracts/L2WrappedBaseToken",
     "l1-contracts/MessageRoot",
-    "l1-contracts/UpgradeableBeacon",
     "l2-contracts/RollupL2DAValidator",
     "l2-contracts/ValidiumL2DAValidator",
     "system-contracts/AccountCodeStorage",
@@ -143,14 +148,12 @@ const EXPECTED_BYTECODES: [&str; 46] = [
     "system-contracts/KnownCodesStorage",
     "system-contracts/L1Messenger",
     "system-contracts/L2BaseToken",
-    "system-contracts/L2GatewayUpgrade",
     "system-contracts/L2GenesisUpgrade",
     "system-contracts/MsgValueSimulator",
     "system-contracts/NonceHolder",
     "system-contracts/PubdataChunkPublisher",
     "system-contracts/SloadContract",
     "system-contracts/SystemContext",
-    "system-contracts/TransparentUpgradeableProxy",
 ];
 
 impl ProposedUpgrade {
@@ -188,16 +191,16 @@ impl ProposedUpgrade {
             result.report_error("Invalid paymaster");
         }
         if tx.nonce != U256::from(expected_version.minor) {
-            result.report_error("Minor protocol version mismatch");
+            result.report_error(&format!(
+                "Minor protocol version mismatch: {} vs {} ",
+                tx.nonce, expected_version.minor
+            ));
         }
         if tx.value != U256::ZERO {
             result.report_error("Invalid value");
         }
         if tx.reserved != [U256::ZERO; 4] {
             result.report_error("Invalid reserved");
-        }
-        if !tx.data.is_empty() {
-            result.report_error("Invalid data");
         }
         if !tx.signature.is_empty() {
             result.report_error("Invalid signature");
@@ -258,6 +261,15 @@ impl ProposedUpgrade {
                 expected_bytecodes
             ));
         }
+        // Check calldata.
+        let calldata = forceDeployOnAddressesCall::abi_decode(&tx.data, true).unwrap();
+        let expected_deployments = expected_force_deployments();
+        verify_force_deployments(
+            &calldata._deployParams,
+            &expected_deployments,
+            verifiers,
+            result,
+        )?;
 
         Ok(())
     }
@@ -274,7 +286,8 @@ impl ProposedUpgrade {
         let initial_error_count = result.errors;
 
         self.verify_transaction(verifiers, result, expected_version, bytecodes_supplier_addr)
-            .await?;
+            .await
+            .context("upgrade tx")?;
 
         result.expect_zk_bytecode(verifiers, &self.bootloaderHash, "proved_batch.yul");
         result.expect_zk_bytecode(
@@ -305,8 +318,9 @@ impl ProposedUpgrade {
             result.report_error("l1ContractsUpgradeCalldata is not empty");
         }
 
-        let post_upgrade_calldata = PostUpgradeCalldata::parse(&self.postUpgradeCalldata)?;
-        post_upgrade_calldata.verify(verifiers, result).await?;
+        if self.postUpgradeCalldata.len() != 0 {
+            result.report_error("Expected empty post upgrade calldata");
+        }
 
         if self.upgradeTimestamp != U256::default() {
             result.report_error("Upgrade timestamp must be zero");